requirements.txt
flask==2.3.3
flask-mongoengine==1.0.0
mongoengine==0.27.0
pymongo==4.6.0
python-dotenv==1.0.0
pinecone-client==3.0.0
langchain==0.1.0
langchain-community==0.0.10
langchain-huggingface==0.0.1
langchain-pinecone==0.0.1
langchain-google-genai==0.0.5
sentence-transformers==2.2.2
pypdf==3.17.4

